Update README.md
The connection between the Arduino Nano and the Pi Pico worked fine via SPI, but there is a problem with communication between the Pi Zero and the Arduino Nano on the same bus. Now the communication will be via USB cable between these boards. This change has been applied to the scheme.