Update keypad.py
A method has been added which, if the keyboard has been connected to a good range of pins, but the connections in this range to the keypad may be mixed, detects (user interaction needed) which pins should be assigned as rows, which as columns, and sorts them accordingly, so that when the key [1] is pressed, the method reads correctly that the key [1] has just been pressed, and so on.

These settings are saved as a picle. The next time a class object is instantiated, this one checks to see if this settings file exists. If so, it loads it, if not, it uses the default pin layout.